Use float to instead of uint16_t in global variable

+++ Program/Program_System/QCopterFC.c
@@ -104,8 +104,8 @@ int main(void)
 			printf("!WARNING: NO SIGNAL!");
 
 		} else {
-			printf("Roll%d,Pitch%d,Yaw%d,CH1 %u(%d),CH2 %u(%d),CH3 %u(%d),CH4 %u(%d),CH5 %u()\r\n",
-			       (int)AngE.Roll, (int)AngE.Pitch, (int)AngE.Yaw,
+			printf("Roll%f,Pitch%f,Yaw%f,CH1 %f(%f),CH2 %f(%f),CH3 %f(%f),CH4 %f(%f),CH5 %f()\r\n",
+			       AngE.Roll, AngE.Pitch, AngE.Yaw,
 			       global_var[PWM1_CCR].param, global_var[RC_EXP_ROLL].param,
 			       global_var[PWM2_CCR].param, global_var[RC_EXP_PITCH].param,
 			       global_var[PWM3_CCR].param, global_var[RC_EXP_THR].param,
